pyproject.toml
[project]
name = "sigilfund"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-migrate>=4.1.0",
    "flask-pydantic>=0.13.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pycoingecko>=3.2.0",
    "pydantic>=2.11.2",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.398",
    "ruff>=0.11.4",
]
